+++ asf_search/CMR/datasets.py
@@ -1994,6 +1994,24 @@ collections_by_processing_level = {
 
 # Helper Methods
 
+def get_nisar_collection_by_maturity(maturities: List[str]):
+    collections = {
+        'BETA': [],
+        'PROVISIONAL': [],
+        'VALIDATED': []
+    }
+
+    for short_name, maturity_collection in dataset_collections['NISAR'].items():
+        if 'BETA' in short_name:
+            collections['BETA'].extend(maturity_collection)
+        elif 'PROVISIONAL' in short_name:
+            collections['PROVISIONAL'].extend(maturity_collection)
+        elif short_name.endswith('V1'):
+            collections['VALIDATED'].extend(maturity_collection)
+    try:
+        return sum([collections[maturity.upper()] for maturity in maturities],[])
+    except KeyError:
+        raise ValueError(f'Invalid maturity option provided valid options are: {", ".join(collections.keys())}' )
 
 def get_concept_id_alias(param_list: List[str], collections_dict: dict) -> List[str]:
     """